Depreciate org-gcal-sync via run-with-idle-timer and use hooks
Previously "org-gcal-sync" was run inside google-calendar/org-gcal-update
function. All this function did was call "org-gcal-sync" and "org-gcal-fetch".
Considering that org-gcal package upstream has made it clear than
"org-gcal-sync" is a two way sync while "org-gcal-fetch" is a one way sync, now
the google calendar is updated with just the "org-gcal-sync".

Previously "org-gcal-sync" was called using "run-with-idle-timer" as per
discussion in original PR. Now it is called when "org-agenda" is called and when
"google-calendar/calfw-view" is called. This makes more contextual sense and
reduces minibuffer clutter during updates.